hackathon_1/hackathon_1/main.py
from dotenv import load_dotenv
import os
import tweepy

load_dotenv()
api_key = os.getenv('API_KEY')
api_secret = os.getenv('API_SECRET')
access_token = os.getenv('ACCESS_TOKEN')
access_secret = os.getenv('ACCESS_SECRET')
XAuth = tweepy.Client(consumer_key=api_key, consumer_secret=api_secret, access_token=access_token, access_token_secret=access_secret)
XAuth.create_tweet(text="tweepytest")